Move adapters to adapters folder
