+++ dexVolumes/index.ts
@@ -1,26 +1,26 @@
-import mooniswap from "./mooniswap";
-import balancer from "./balancer";
-import bancor from "./bancor";
-import carthage from "./carthage";
-import champagneswap from "./champagneswap";
-import curve from "./curve";
-import dodo from "./dodo";
-import katana from "./katana";
-import klayswap from "./klayswap";
-import osmosis from "./osmosis";
-import pancakeswap from "./pancakeswap";
-import quickswap from "./quickswap";
-import raydium from "./raydium";
-import saros from "./saros";
-import serum from "./serum";
-import soulswap from "./soulswap";
-import spiritswap from "./spiritswap";
-import spookyswap from "./spookyswap";
-import sushiswap from "./sushiswap";
-import terraswap from "./terraswap";
-import traderjoe from "./traderjoe";
-import uniswap from "./uniswap";
-import yieldfields from "./yieldfields";
+import mooniswap from "./adapters/mooniswap";
+import balancer from "./adapters/balancer";
+import bancor from "./adapters/bancor";
+import carthage from "./adapters/carthage";
+import champagneswap from "./adapters/champagneswap";
+import curve from "./adapters/curve";
+import dodo from "./adapters/dodo";
+import katana from "./adapters/katana";
+import klayswap from "./adapters/klayswap";
+import osmosis from "./adapters/osmosis";
+import pancakeswap from "./adapters/pancakeswap";
+import quickswap from "./adapters/quickswap";
+import raydium from "./adapters/raydium";
+import saros from "./adapters/saros";
+import serum from "./adapters/serum";
+import soulswap from "./adapters/soulswap";
+import spiritswap from "./adapters/spiritswap";
+import spookyswap from "./adapters/spookyswap";
+import sushiswap from "./adapters/sushiswap";
+import terraswap from "./adapters/terraswap";
+import traderjoe from "./adapters/traderjoe";
+import uniswap from "./adapters/uniswap";
+import yieldfields from "./adapters/yieldfields";
 
 export default {
   mooniswap,
